refactor: isolate workspace snapshot process diagnostics
